fix filesystemctx to use original getallworkers as default
